requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
access=1.1.3=pyhd3eb1b0_0
affine=2.3.0=pyhd3eb1b0_0
amply=0.1.4=py_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
arviz=0.11.4=pyhd8ed1ab_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.6.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.10.0=pyh06a4308_0
bleach=4.1.0=pyhd3eb1b0_0
blosc=1.21.0=h9c3ff4c_0
boost-cpp=1.74.0=h312852a_4
bottleneck=1.3.2=py39hdd57654_1
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h27cfd23_1003
bs4=4.10.0=hd3eb1b0_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.8=ha878542_0
cachetools=5.0.0=pyhd8ed1ab_0
cairo=1.16.0=h6cf1ce9_1008
certifi=2021.10.8=py39hf3d152e_1
cffi=1.15.0=py39hd667e15_1
cfitsio=3.470=hb418390_7
cftime=1.5.2=py39hce5d2b2_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=7.1.2=pyhd3eb1b0_0
click-plugins=1.1.1=pyhd3eb1b0_0
cligj=0.7.2=py39h06a4308_0
coincbc=2.10.5=hcee13e7_1
coverage=6.2=py39h7f8727e_0
cryptography=36.0.0=py39h9ce1e76_0
curl=7.80.0=h7f8727e_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
deprecation=2.1.0=pyhd3eb1b0_0
dill=0.3.4=pyhd8ed1ab_0
docutils=0.18.1=py39h06a4308_2
entrypoints=0.3=py39h06a4308_0
esda=2.4.1=pyhd3eb1b0_0
expat=2.4.4=h295c915_2
fastprogress=1.0.2=pyhd8ed1ab_0
filelock=3.5.1=pyhd8ed1ab_0
fiona=1.8.20=py39h494ca2c_0
flask=2.0.2=pypi_0
flask-sock=0.4.0=pypi_0
fontconfig=2.13.1=hba837de_1005
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
freexl=1.0.6=h27cfd23_0
gdal=3.2.2=py39h409cc32_8
geopandas=0.9.0=py_1
geopandas-base=0.9.0=py_1
geos=3.9.1=h295c915_0
geotiff=1.6.0=h4f31c25_6
gettext=0.21.0=hf68c758_0
giddy=2.3.3=py_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gmpy2=2.0.8=py39h8083e48_3
hdf4=4.2.15=h10796ff_3
hdf5=1.10.6=nompi_h6a2412b_1114
icu=68.1=h2531618_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.8.2=py39h06a4308_0
importlib_metadata=4.8.2=hd3eb1b0_0
inequality=1.0.0=pyhd3eb1b0_0
iniconfig=1.1.1=pyhd3eb1b0_0
ipykernel=6.4.1=py39h06a4308_1
ipython=7.31.1=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
itsdangerous=2.0.1=pypi_0
jbig=2.1=hdba287a_0
jedi=0.18.1=py39h06a4308_1
jinja2=3.0.2=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
json-c=0.15=h98cffda_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.1=py39h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kealib=1.4.14=hb50703a_1
kiwisolver=1.3.2=py39h295c915_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=11_linux64_openblas
libcblas=3.9.0=11_linux64_openblas
libcurl=7.80.0=h0b77cf5_0
libdap4=3.20.6=hd7c4107_2
libdeflate=1.7=h7f98852_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.1.0=hc902ee8_8
libgdal=3.2.2=h07ae0cd_8
libgfortran-ng=9.3.0=ha5ec8a7_17
libgfortran5=9.3.0=ha5ec8a7_17
libglib=2.68.3=h3e27bee_0
libiconv=1.16=h516909a_0
libkml=1.3.0=h238a007_1014
liblapack=3.9.0=11_linux64_openblas
libllvm11=11.1.0=h3826bc1_0
libnetcdf=4.8.1=nompi_hcd642e3_100
libnghttp2=1.46.0=hce63b2e_0
libopenblas=0.3.17=pthreads_h8fe5266_1
libpng=1.6.37=hbc83047_0
libpq=13.3=hd57d9b9_0
libpysal=4.5.1=pyhd3eb1b0_0
librttopo=1.1.0=h1185371_6
libsodium=1.0.18=h7b6447c_0
libspatialindex=1.9.3=h2531618_0
libspatialite=5.0.1=h8694cbe_5
libssh2=1.9.0=h1ba5d50_1
libstdcxx-ng=11.2.0=he4da1e4_12
libtiff=4.3.0=hf544144_1
libuuid=2.32.1=h7f98852_1000
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h72842e0_0
libzip=1.8.0=h4de3113_0
llvm-openmp=12.0.1=h4bd325d_1
llvmlite=0.37.0=py39h295c915_1
lxml=4.7.1=pypi_0
lz4-c=1.9.3=h295c915_1
lzo=2.10=h7b6447c_2
mapclassify=2.4.3=pyhd3eb1b0_0
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py39h27cfd23_0
matplotlib-base=3.5.0=py39h3ed280b_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mgwr=2.1.2=py_0
mistune=0.8.4=py39h27cfd23_1000
mkl=2022.0.1=h8d4b97c_803
mkl-service=2.4.0=py39h404a4ab_0
mock=4.0.3=pyhd3eb1b0_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.2.1=py39h06a4308_0
munch=2.5.0=pyhd3eb1b0_0
munkres=1.1.4=py_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.3.0=py39h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
netcdf4=1.5.7=py39ha0f2276_1
networkx=2.6.3=pyhd3eb1b0_0
nose=1.3.7=pyhd3eb1b0_1008
notebook=6.4.8=py39h06a4308_0
numba=0.54.1=py39h51133e4_0
numexpr=2.7.3=py39hde0f152_0
numpy=1.20.3=py39hdbf815f_1
olefile=0.46=pyhd3eb1b0_0
opencv-contrib-python=4.5.3.56=pypi_0
openjpeg=2.4.0=h3ad879b_0
openssl=1.1.1l=h7f98852_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py39h8c16a72_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
patsy=0.5.2=py39h06a4308_1
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
pixman=0.40.0=h7f8727e_1
pluggy=1.0.0=py39h06a4308_0
pointpats=2.2.0=py_0
poppler=21.03.0=h93df280_0
poppler-data=0.4.11=h06a4308_0
postgresql=13.3=h2510834_0
proj=8.0.1=h1217e81_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pulp=2.5.0=py39hf3d152e_0
py=1.11.0=pyhd3eb1b0_0
pycodestyle=2.8.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygeos=0.10.1=py39ha61afbd_0
pygments=2.11.2=pyhd3eb1b0_0
pymc3=3.11.2=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyproj=3.1.0=py39ha9a7ae0_3
pyrsistent=0.18.0=py39heee7806_0
pysal=2.5.0=pyhd8ed1ab_0
pysocks=1.7.1=py39h06a4308_0
pytables=3.6.1=py39h77479fe_1
pytest=6.2.5=py39h06a4308_2
pytest-cov=3.0.0=pyhd3eb1b0_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.0=py_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py39h7f8727e_1
pyzmq=22.3.0=py39h295c915_2
quantecon=0.5.0=py39h06a4308_0
rasterio=1.2.6=py39hbc9cc8e_1
rasterstats=0.14.0=py_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
rtree=0.9.7=py39h06a4308_1
rvlib=0.0.5=py39hce5d2b2_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.6.3=py39hee8e79c_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
segregation=2.1.0=pyhd8ed1ab_0
selenium=3.141.0=py39h27cfd23_1000
semver=2.13.0=pyh9f0ad1d_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py39h06a4308_0
shapely=1.7.1=py39ha61afbd_5
simple-websocket=0.5.0=pypi_0
simplejson=3.17.6=py39h7f8727e_0
six=1.16.0=pyhd3eb1b0_1
snuggs=1.4.7=pyhd3eb1b0_0
soupsieve=2.3.1=pyhd3eb1b0_0
spaghetti=1.6.2=pyhd3eb1b0_0
spglm=1.0.8=py_0
spint=1.0.7=py_0
splot=1.1.3=py_0
spopt=0.1.2=pyhd8ed1ab_0
spreg=1.2.4=pyhd3eb1b0_0
spvcm=0.3.0=py_0
sqlite=3.37.2=hc218d9a_0
statsmodels=0.12.2=py39h27cfd23_0
sympy=1.9=py39h06a4308_0
tbb=2021.5.0=hd09550d_0
terminado=0.13.1=py39h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
theano-pymc=1.1.2=py39h51133e4_0
threadpoolctl=2.2.0=pyh0d69192_0
tiledb=2.3.4=he87e0bf_0
tk=8.6.11=h1ccaba5_0
tobler=0.8.2=pyhd8ed1ab_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py39h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
turbo-flask=0.8.0=pypi_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzcode=2021a=h7f98852_2
tzdata=2021e=hda174b7_0
urllib3=1.25.11=py_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
werkzeug=2.0.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39h06a4308_0
xarray=0.21.1=pyhd8ed1ab_0
xerces-c=3.2.3=h9d8b166_2
xlrd=2.0.1=pyhd3eb1b0_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h27cfd23_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.5.0=ha4553b6_1
